wikicomic_vercel_ready/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
wikipedia==1.4.0
groq==0.4.2
requests==2.31.0
Pillow==10.1.0
python-dotenv==1.0.0
google-generativeai==0.3.2
pymongo==4.6.0
gridfs==0.0.3
bson==0.5.10